+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### 2.9.0-beta.1 
 
-> [!IMPORTANT]
+> [!Breaking Changes] 
 > [Android] When a method returns `StylePropertyValue` the property values will now be typed rather than a string.
+> `PointAnnotation.iconImageCrossFade` has been deprecated and setting a value to it will not have any impact. Use `PointAnnotationManager.iconImageCrossFadeTransition` instead.
 
 * In this release we fixed a bug in our Android conversion code where the property values in `StylePropertyValue` were being returned as strings rather than their actual type. This fix will cause a behavioral change in the return value of the following methods on Android:
   * `getStyleImportConfigProperties`, `getStyleImportConfigProperty`, `getStyleLayerProperty`, `getStyleSourceProperty`, `getStyleTerrainProperty`, `getStyleLightProperty`.