+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-### main
+### 2.9.0-rc.1
+
+* Update MapboxMaps to v11.13.0-rc.1
 
 ### 2.9.0-beta.1